training-portal/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.1
django-taggit==3.0.0
sqlparse==0.4.3
tzdata==2022.2
